claude-agent-sdk-demo/pyproject.toml
[project]
name = "claude-agent-sdk-demo"
version = "0.1.0"
description = "Claude Agent SDK integration with LangSmith tracing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langsmith[claude-agent-sdk]>=0.4.38",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]
